Finalised Assignment 3, Semester 1
Voltage readings are obtained from an electrical substation once every hour for six hours (so there are six readings).  Write a program to perform the following checks on the substation:

i)  display all voltages that differ from the average by more than 10% of the average.
ii) display all pairs of consecutive hours where the change from the voltage at one hour to the next is greater than 15% of the average.